requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
filelock==3.24.3
flask==3.1.1
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ImageHash==4.3.2
joblib==1.5.3
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
plyer==2.1.0
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
pystray==0.19.5
PyWavelets==1.9.0
pywin32-ctypes==0.2.3
PyYAML==6.0.3
regex==2026.2.19
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
watchdog==6.0.0
